fix: improve driver activity data parsing

+++ append_card_places.go
@@ -45,7 +45,7 @@ func AppendPlaceRecord(dst []byte, rec *cardv1.Places_Record) ([]byte, error) {
 	dst = append(dst, byte(countryProtocol)) // 1 byte
 
 	dst = binary.BigEndian.AppendUint16(dst, uint16(rec.GetDailyWorkPeriodRegion())) // 2 bytes
-	dst = appendOdometer(dst, rec.GetVehicleOdometerKm())                            // 3 bytes
+	dst = appendOdometer(dst, uint32(rec.GetVehicleOdometerKm()))                            // 3 bytes
 	dst = append(dst, byte(rec.GetReservedByte()))                                   // 1 byte reserved (preserved)
 	return dst, nil
 }
